requirements.txt
Flask==3.0.3
imageio==2.33.0
lightning==2.5.5
matplotlib==3.6.3
numpy==1.24.2
Pillow==10.3.0
pynrrd==1.0.0
scikit_learn==1.3.2
torch==2.1.2
torchmetrics==1.2.1
torchvision==0.16.2
waitress==3.0.0
openai==0.27.8
